Re-consolidate pkg main (❌ App Engine compat)
