--- cmd/go_httpbin/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/mccutchen/go-httpbin/v2/cmd/maincmd"
-
-func main() {
-	maincmd.Main()
-}
